Add the new login page
Note that the changes to the _signin and _signup partials are
very slight - the submit button text changes in each, and the
order of the name and email entry fields are swapped in _signup.

Another thing to be aware of is that the form of the sentences
that include the reasons why a sign-in / sign-up is required
haven't been finalized.  (See the discussion at
mysociety/alaveteli#1092.)