Add stub text for the time being to unbreak this.
We need to decide if we're previewing the email which will be sent to
the authority, or the way this request will appear on the site. If we're previewing the appearance on the site, we need some bit of text for the case where you're not logged in to indicate that this is where your name will go.